refactor(kitchenpos): move to common package
- move Profanities, DisplayedName, Price to common package
