+++ README.md
@@ -158,12 +158,12 @@
 
 ### 메뉴
 
-- `MenuGroup`은 식별자와 이름을 가진다.
+- `MenuGroup`은 식별자와 `Name`을 가진다.
 - `Menu`는 식별자와 `DisplayedName`, `Price`, `MenuProducts`를 가진다.
 - `Menu`는 특정 `MenuGroup`에 속한다.
 - `Menu`의 `Price`은 `MenuProducts`의 금액의 합보다 적거나 같아야 한다.
 - `Menu`의 `Price`이 `MenuProducts`의 금액의 합보다 크면 `NotDisplayedMenu`가 된다.
-- `MenuProduct`는 `Price`과 수량을 가진다.
+- `MenuProduct`는 `Price`과 `Quantity`을 가진다.
 
 ### 매장 주문
