+++ src/main/java/kitchenpos/common/domain/Profanities.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.common.domain;
 
 public interface Profanities {
 	boolean contains(String name);
