+++ src/main/java/kitchenpos/products/tobe/domain/Product.java
@@ -6,10 +6,11 @@ import java.util.UUID;
 import javax.persistence.Column;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
+
+import kitchenpos.common.domain.DisplayedName;
+import kitchenpos.common.domain.Price;
 
 @Table(name = "product")
 @Entity(name = "TobeProduct")
